% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ActuarialTable.R
\docType{class}
\name{ActuarialTable-class}
\alias{ActuarialTable}
\alias{ActuarialTable-class}
\title{ActuarialTable}
\description{
An S4 class to represent an actuarial table
}
\section{Slots}{

\describe{
\item{\code{i}}{a length one or length n numeric vector of interest rates}
}}

