% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/apv.R
\docType{methods}
\name{apv,T_x-method}
\alias{apv,T_x-method}
\title{apv}
\usage{
\S4method{apv}{T_x}(object, t = NULL, m = 0)
}
\arguments{
\item{object}{object of class LifeTable}

\item{x_}{x}

\item{t_}{t}
}
\description{
apv
}
\examples{
apv(new("T_x"), t = 3)
}

