% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Benefit.R
\docType{class}
\name{Benefit-class}
\alias{Benefit}
\alias{Benefit-class}
\title{Benefit}
\description{
virtual class inherited by all types of benefits. (e.g. BenefitDeath and
BenefitAnnuity inherit the Benefit class)
}

