% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/benefit.R
\docType{class}
\name{BenefitAnnuity-class}
\alias{BenefitAnnuity}
\alias{BenefitAnnuity-class}
\title{BenefitAnnuity}
\description{
Annuity Benefit values.  This is meant to be used with the \code{Insuree}
class
}
\section{Slots}{

\describe{
\item{\code{t}}{time (from present moment) that the benefit will be payed}

\item{\code{value}}{numeric vector defining value of over the t time}
}}

