% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Benefit.R
\docType{class}
\name{BenefitDeath-class}
\alias{BenefitDeath}
\alias{BenefitDeath-class}
\title{BenefitDeath}
\description{
Death Benefit values.  This is meant to be used with the \code{Insuree}
class
}
\section{Slots}{

\describe{
\item{\code{t}}{time over which death benefit value is applicable}

\item{\code{value}}{numeric vector defining value of death benefit over t time}
}}

