% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ActuarialTable.R
\name{check_ActuarialTable}
\alias{check_ActuarialTable}
\title{check_ActuarialTable}
\usage{
check_ActuarialTable(object)
}
\arguments{
\item{object}{object of class ActuarialTable

function to check validity of LifeTable S4 class constructor}
}
\description{
check_ActuarialTable
}

