% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/benefit.R
\name{check_BenefitAnnuity}
\alias{check_BenefitAnnuity}
\title{check_BenefitAnnuity}
\usage{
check_BenefitAnnuity(object)
}
\arguments{
\item{object}{object of class \code{BenefitAnnuity}

function to check validity of DeathBenefit S4 class constructor}
}
\description{
check_BenefitAnnuity
}

