% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/benefit.R
\name{check_BenefitDeath}
\alias{check_BenefitDeath}
\title{check_BenefitDeath}
\usage{
check_BenefitDeath(object)
}
\arguments{
\item{object}{object of class \code{BenefitDeath}

function to check validity of DeathBenefit S4 class constructor}
}
\description{
check_BenefitDeath
}

