% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Insuree.R
\name{check_Insuree}
\alias{check_Insuree}
\title{check_Insuree}
\usage{
check_Insuree(object)
}
\arguments{
\item{object}{object of class \code{Insuree}

function to check validity of Insuree S4 class constructor}
}
\description{
check_Insuree
}

