% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Interest.R
\name{check_Interest}
\alias{check_Interest}
\title{check_Interest}
\usage{
check_Interest(object)
}
\arguments{
\item{object}{object of class \code{Interest}

function to check validity of DeathBenefit S4 class constructor}
}
\description{
check_Interest
}

