% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Life.R
\name{check_Life}
\alias{check_Life}
\title{check_Life}
\usage{
check_Life(object)
}
\arguments{
\item{object}{object of class \code{Life}

function to check validity of Insuree S4 class constructor}
}
\description{
check_Life
}

