% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LifeTable.R
\name{check_LifeTable}
\alias{check_LifeTable}
\title{check_LifeTable}
\usage{
check_LifeTable(object)
}
\arguments{
\item{object}{object of class \code{Insuree}

function to check validity of LifeTable S4 class constructor}
}
\description{
check_LifeTable
}

