% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pool.R
\name{check_Pool}
\alias{check_Pool}
\title{check_Pool}
\usage{
check_Pool(object)
}
\arguments{
\item{object}{object of class \code{Pool}

function to check validity of \code{Pool} S4 class}
}
\description{
check_Pool
}

