% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/discount.R
\docType{methods}
\name{discount,BenefitDeath-method}
\alias{discount,BenefitDeath-method}
\title{discount}
\usage{
\S4method{discount}{BenefitDeath}(object, tod)
}
\arguments{
\item{object}{object of class DeathBenefit}

\item{tod}{time of death}
}
\description{
Discounts a \code{DeathBenefit} object
}
\examples{
discount(object = BenefitDeath(), tod = 2)
discount(object = BenefitDeath(), tod = 10)
}

