% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/discount.R
\docType{methods}
\name{discount,Interest,ANY-method}
\alias{discount,Interest,ANY-method}
\title{discount}
\usage{
\S4method{discount}{Interest,ANY}(interest, benefit = NULL, benefit_t)
}
\arguments{
\item{interest}{object of class \code{Interest}}

\item{benefit_t}{time of benefit payment}
}
\description{
Finds the discount applicable to a future time of death
}
\examples{
discount(Interest(), benefit_t = 2.1)
discount(Interest(), benefit_t = 1.5)
discount(Interest(), benefit_t = 0.5)

i <- Interest(t = rep(1, times = 5), rate = c(0.03, 0.06, 0.05, 0.07, 0.08))
}

