% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/discount.R
\docType{methods}
\name{discount,Interest,BenefitAnnuity-method}
\alias{discount,Interest,BenefitAnnuity-method}
\title{discount}
\usage{
\S4method{discount}{Interest,BenefitAnnuity}(interest, benefit, tod)
}
\arguments{
\item{interest}{object of class \code{Interest}}

\item{benefit}{object of class \code{Benefit}}

\item{tod}{the time of the benefit payment}
}
\description{
discount a \code{BenefitAnnuity} object
}
\examples{
discount(Interest(), BenefitAnnuity(), tod = 2.1)
discount(Interest(), BenefitAnnuity(), tod = NA)
discount(Interest(), BenefitAnnuity(), tod = 10)
discount(Interest(), BenefitAnnuity(t = c(1, 1, 1), value = c(3, 2, 2)), tod = 3)
discount(Interest(), BenefitAnnuity(t = c(1, 1, 1), value = c(3, 2, 2)), tod = 0.5)
}

