% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/discount.R
\docType{methods}
\name{discount,Interest,BenefitDeath-method}
\alias{discount,Interest,BenefitDeath-method}
\title{discount}
\usage{
\S4method{discount}{Interest,BenefitDeath}(interest, benefit, tod)
}
\arguments{
\item{interest}{object of class \code{Interest}}

\item{benefit}{object of class \code{DeathBenefit}}

\item{tod}{time of death}
}
\description{
Discounts a \code{DeathBenefit} object
}
\examples{
discount(benefit = BenefitDeath(), interest = Interest(), tod = 2)
discount(benefit = BenefitDeath(), interest = Interest(), tod = NA)
discount(benefit = BenefitDeath(t = 10,
                                value = 1000),
         interest = Interest(t = rep(1, times = 10),
                             rate = rep(c(0.03, 0.04), times = 5)),
         tod = 8)
}

