% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/discount.R
\name{discount}
\alias{discount}
\title{discount}
\usage{
discount(object, tod)
}
\arguments{
\item{object}{discount objects defining the benefit payment values and times}

\item{tod}{time of death}
}
\description{
discounts a life contingent benefit
}

