% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{discount_death}
\alias{discount_death}
\title{discount_death}
\usage{
discount_death(object, death_time = NA)
}
\arguments{
\item{object}{object of class Insuree}

\item{death_time}{}
}
\description{
function to discount single death benefit
}
\examples{
discount_death(Insuree(), death_time = 2.7)
}

