% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/e-x.R
\name{e_x}
\alias{e_x}
\title{e_x}
\usage{
e_x(life_table, x, t = NULL)
}
\arguments{
\item{life_table}{object of class life_table}

\item{x}{x}

\item{t}{t}
}
\description{
calculates the kurtate future life expectancy
}
\examples{
my_table <- data.frame("x" = 0:9,
                       "q_x" = seq(0.05, 0.14, by = 0.01)
                      )

my_table <- life_table(my_table, x = "x", q_x = "q_x")
e_x(my_table, x = 3)
}

