% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/expected.R
\docType{methods}
\name{expected,Insuree-method}
\alias{expected,Insuree-method}
\title{expected}
\usage{
\S4method{expected}{Insuree}(object)
}
\arguments{
\item{object}{object of class Insuree}

\item{t_}{t}

\item{m_}{m}
}
\description{
expected present value (discounted for interest and mortality)
of a single death benefit of 1.
}
\examples{
# 3 year term insurance with benefit of 4
expected(Insuree(t_ = 3, benefit = 4))
# 3 year term insurance effective in 1 year
expected(Insuree(benefit = 1, x_ = 0, m_ = 1, t_ = 3))
}

