% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/expected.R
\docType{methods}
\name{expected,LifeTable-method}
\alias{expected,LifeTable-method}
\title{expected}
\usage{
\S4method{expected}{LifeTable}(object, x_ = object@x[1], t_ = NULL,
  m_ = 0)
}
\arguments{
\item{object}{object of class LifeTable}

\item{x_}{x_}

\item{t_}{t_}

\item{m_}{m_}
}
\description{
expected number of years lived during term period where term period
is defined as the period between x_ + m_ and t_.
}
\examples{
expected(LifeTable(), x_ = 2, t_ = 3, m_ = 0)
}

