% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/expected.R
\docType{methods}
\name{expected,Y_x-method}
\alias{expected,Y_x-method}
\title{expected}
\usage{
\S4method{expected}{Y_x}(object, t_ = NULL, m_ = 0, payable = "end")
}
\arguments{
\item{object}{object of class Z_x}

\item{t_}{t}

\item{m_}{m}

\item{payable}{character string indicating the time the annuity
payments are made.  Can either be "beg" or "end".  Default value
set to "end".}
}
\description{
expected value of \code{Y_x}
}
\examples{
expected(new("Y_x"), t_ = 3)
}

