% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/expected.R
\docType{methods}
\name{expected,Z_x-method}
\alias{expected,Z_x-method}
\title{expected}
\usage{
\S4method{expected}{Z_x}(object, t_ = max(object@x) - object@x_, m_ = 0)
}
\arguments{
\item{object}{object of class Z_x}

\item{t_}{t}

\item{m_}{m}
}
\description{
expected present value of an insurance payment \code{Z_x}.  \code{Z_x}
is discounted for interest and mortality.
}
\examples{
# 3 year term insurance with a payable amount or 4
expected(Z_x(), t_ = 3)
# 3 year term insurance effective in 1 year
expected(Z_x(benefit = 1, x_ = 0), t_ = 3, m_ = 1)
}

