% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/expected.R
\name{expected}
\alias{expected}
\title{expected}
\usage{
expected(object, x_, t_ = NULL, m_ = 0)
}
\arguments{
\item{object}{object of class LifeTable}

\item{x_}{x_}

\item{t_}{t_}

\item{m_}{m_}
}
\description{
calculates the actuarial present value of an individual at
birthday \code{x}.
}

