% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.R
\name{hist.rpv_Insuree}
\alias{hist.rpv_Insuree}
\title{hist.rpv_Insuree}
\usage{
\method{hist}{rpv_Insuree}(object, ...)
}
\arguments{
\item{object}{rpv_life object}

\item{...}{other arguments}
}
\description{
plot a histogram of the present value from a rpv_life() simulation
}
\examples{
test <- rpv(object = Insuree(),
            n = 500,
            interest = 0.04)
hist(test)
}

