% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.R
\name{hist.rpv_life}
\alias{hist.rpv_life}
\title{hist.rpv_life}
\usage{
\method{hist}{rpv_life}(object)
}
\arguments{
\item{object}{rpv_life object}
}
\description{
plot a histogram of the present value from a rpv_life() simulation
}
\examples{
test <- rpv_life(object = Insuree(),
                 n = 500,
                 interest = 0.04)
hist(test)
}

