% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.R
\name{hist.rpv_Pool}
\alias{hist.rpv_Pool}
\title{hist.rpv_Pool}
\usage{
\method{hist}{rpv_Pool}(object, ...)
}
\arguments{
\item{object}{rpv_Pool object}

\item{...}{other arguments}
}
\description{
plot a histogram of the present value from a rpv() simulation of the
\code{Pool} class
}
\examples{
test <- rpv(object = Pool(),
            n = 500,
            interest = 0.04)
hist(test)
}

