% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{index_x}
\alias{index_x}
\title{index_x}
\usage{
index_x(x, x_, m_t_ = 1)
}
\arguments{
\item{x}{vector of ascending numerics}

\item{x_}{the index of the first x value to be returned.}

\item{m_t_}{the index of the last x value to be returned.  This is oftern m_ + t_, m_
or t_.}
}
\description{
Returns the x values for an actuarial table
}
\examples{
my_x <- 0:9
index_x(my_x, x_ = 2, m_t_ = 3)
index_x(my_x, x_ = 2.4, m_t_ = 3)
index_x(my_x, x_ = 2.4, m_t_ = 0)
}
\keyword{internal}

