% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Interest.R
\docType{class}
\name{Interest-class}
\alias{Interest}
\alias{Interest-class}
\title{Interest}
\description{
intrest rate class for discounting benefits
}
\section{Slots}{

\describe{
\item{\code{t}}{time over which the annualized intereste rate is applicable}

\item{\code{rate}}{annualized interest rate}
}}

