% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Life.R
\docType{class}
\name{Life-class}
\alias{Life}
\alias{Life-class}
\title{Life}
\description{
An individual with some kind of life contingent insurance
}
\details{
#@include LifeTable.R
#@include Interest.R
#@include Benefit.R
}
\section{Slots}{

\describe{
\item{\code{x_}}{x value for individual}

\item{\code{t_}}{t value for individual}

\item{\code{m_}}{m value for individual}

\item{\code{lifetable}}{onject of class \code{LifeTable}}

\item{\code{benefit}}{a list of objects of class \code{Benefit}}
}}
\examples{
Life()
}

