% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/life-table.R
\name{life_table}
\alias{life_table}
\title{life_table}
\usage{
life_table(df, x, q_x)
}
\arguments{
\item{df}{data.frame containing x and q_x}

\item{x}{x}

\item{q_x}{q_x}
}
\description{
life_table S3 class constructor
}
\examples{
my_table <- data.frame("x" = 0:9,
                       "q_x" = seq(0.05, 0.14, by = 0.01)
                      )

life_table(my_table, x = "x", q_x = "q_x")
}

