% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LifeTable.R
\docType{class}
\name{LifeTable-class}
\alias{LifeTable}
\alias{LifeTable-class}
\title{LifeTable}
\description{
An S4 class to represent a life table
}
\section{Slots}{

\describe{
\item{\code{x}}{a length n ascending numeric vector}

\item{\code{t}}{a length n numeric vecotr indicating the t_ value
for each x appropriate that corresponds with the q_x element in
the table.  Important for last element and for deferral insurance.}

\item{\code{q_x}}{a length n numeric vector for probability of death at
each element of x}
}}

