\name{lifetable-package}
\alias{lifetable-package}
\alias{lifetable}
\docType{package}
\title{
\packageTitle{lifetable}
}
\description{
\packageDescription{lifetable}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{lifetable}
\packageIndices{lifetable}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{lifetable}

Maintainer: \packageMaintainer{lifetable}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
