% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mean.R
\docType{methods}
\name{mean,T_x-method}
\alias{mean,T_x-method}
\title{mean}
\usage{
\S4method{mean}{T_x}(object, t_ = NULL, m_ = 0)
}
\arguments{
\item{object}{object of class T_x}

\item{t_}{t}

\item{m_}{m}
}
\description{
expected number of complete years for \code{T_x} to survive
}
\examples{
mean(new("T_x"), t_ = 3)
}

