% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mean.R
\docType{methods}
\name{mean,Z_x-method}
\alias{mean,Z_x-method}
\title{mean}
\usage{
\S4method{mean}{Z_x}(object, t_ = NULL, m_ = 0)
}
\arguments{
\item{object}{object of class Z_x}

\item{t_}{t}

\item{m_}{m}
}
\description{
expected number of complete years for \code{Z_x} to survive discounted for interest
}
\examples{
mean(new("Z_x"), t_ = 3)
}

