% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/expected.R
\docType{methods}
\name{mean_life,Insuree-method}
\alias{mean_life,Insuree-method}
\title{mean_life}
\usage{
\S4method{mean_life}{Insuree}(object)
}
\arguments{
\item{object}{object of class \code{Insuree}}

\item{x_}{x_}

\item{t_}{t_}

\item{m_}{m_}
}
\description{
The expected time lived provided an object of class \code{Insuree}
in the \code{object} argument.  Expected time lived only includes times
between x_ + m_ and t_.
}
\examples{
mean_life(Insuree())
mean_life(Insuree(x_ = 2.5, t_ = 3, m_ = 1))
}

