% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/expected.R
\docType{methods}
\name{mean_life,Life-method}
\alias{mean_life,Life-method}
\title{mean_life}
\usage{
\S4method{mean_life}{Life}(object)
}
\arguments{
\item{object}{object of class \code{Life}}
}
\description{
The expected time lived provided an object of class \code{Life}
in the \code{object} argument.  Expected time lived only includes times
between x_ + m_ and t_.
}
\examples{
mean_life(Life())
mean_life(Life(x_ = 2.5, t_ = 3, m_ = 1))
}

