% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/expected.R
\docType{methods}
\name{mean_life,LifeTable-method}
\alias{mean_life,LifeTable-method}
\title{mean_life}
\usage{
\S4method{mean_life}{LifeTable}(object, x_ = object@x[1], t_ = NULL,
  m_ = 0)
}
\arguments{
\item{object}{object of class LifeTable}

\item{x_}{x_}

\item{t_}{t_}

\item{m_}{m_}
}
\description{
The expected time lived provided an object of class \code{LifeTable}
in the \code{object} argument.  Expected time lived only includes times
between x_ + m_ and t_.
}
\examples{
mean_life(LifeTable(), x_ = 2, t_ = 3, m_ = 0)
mean_life(LifeTable(), x_ = 2, t_ = 3, m_ = 1)
mean_life(LifeTable(), x_ = 2.5, t_ = 3, m_ = 1)
}

