% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/expected.R
\name{mean_life}
\alias{mean_life}
\title{mean_life}
\usage{
mean_life(object, ...)
}
\arguments{
\item{object}{object of one of several insuree package classes}

\item{...}{addtitional parameters}
}
\description{
calculates the mean amount of time the \code{object} argument is expected
to live
}

