% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{new_x}
\alias{new_x}
\title{new_x}
\usage{
new_x(x, x_, m_t_)
}
\arguments{
\item{x}{vector of ascending numerics}

\item{x_}{the first x value to be returned.}

\item{m_t_}{the last x value to be returned.  This is oftern m_ + t_, m_
or t_.}
}
\description{
create new x values by specifying new age x_ and
interval m_t_
}
\examples{
new_x(x = 0:5, x_ = 2.3, m_t_ = 1.8)
}

