% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/p-x.R
\docType{methods}
\name{p_x,LifeTable-method}
\alias{p_x,LifeTable-method}
\title{p_x}
\usage{
\S4method{p_x}{LifeTable}(object, t_ = 1, x_ = min(object@x))
}
\arguments{
\item{object}{object of class LifeTable}

\item{t_}{t_}

\item{x_}{x_}
}
\description{
probability of survival for person at birthday x
}
\examples{
p_x(LifeTable(), x_ = 3, t_ = 5) # probability of x = 3 surviving 5 years
}

