% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p-x.R
\name{p_x}
\alias{p_x}
\title{p_x}
\usage{
p_x(object, x_, t_ = 1)
}
\arguments{
\item{object}{object of class LifeTable}

\item{x_}{x_}

\item{t_}{t_}
}
\description{
probability of survival for person at birthday \code{x}
}

