% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.rpv_Insuree}
\alias{plot.rpv_Insuree}
\title{plot.rpv_Insuree}
\usage{
\method{plot}{rpv_Insuree}(object, ...)
}
\arguments{
\item{object}{rpv_life object}

\item{...}{other arguments}
}
\description{
plot the empirical cumulative distribution of
the present value of death benefits from a rpv_life()
simulation
}
\examples{
test <- rpv(object = Insuree(),
            n = 500,
            interest = 0.04)
plot(test)
}

