% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.rpv_Life}
\alias{plot.rpv_Life}
\title{plot.rpv_Life}
\usage{
\method{plot}{rpv_Life}(object, ...)
}
\arguments{
\item{object}{object of class rpv_Life}

\item{...}{other arguments}
}
\description{
plot the empirical cumulative distribution of
the present value of death benefits from a rpv()
simulation
}
\examples{
test <- rpv(object = Life(benefit = list(BenefitDeath())),
            n = 500,
            interest = Interest(t = 5, rate = 0.04))
plot(test)
}

