% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Pool.R
\docType{class}
\name{Pool-class}
\alias{Pool}
\alias{Pool-class}
\title{Pool}
\description{
S4 class for a group of \code{Life} objects
}
\section{Slots}{

\describe{
\item{\code{insurees}}{list of objects of class \code{Life}}
}}

