% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/q-x.R
\docType{methods}
\name{q_x,LifeTable-method}
\alias{q_x,LifeTable-method}
\title{q_x}
\usage{
\S4method{q_x}{LifeTable}(object, x_, t_ = 1, m_ = 0)
}
\arguments{
\item{x_}{x_}

\item{t_}{t_}

\item{m_}{m_}

\item{life_table}{object of class life_table}
}
\description{
probability of death for person at birthday x
}
\examples{
q_x(new("LifeTable"), x = 3, t = 5) # probability of x = 3 surviving 5 years
q_x(new("LifeTable"), x = 2, t = 4, m = 1) # probability of x = 2 surviving 1 year and
}

