% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/q-x.R
\name{q_x}
\alias{q_x}
\title{q_x}
\usage{
q_x(life_table, x, t = 1, m = 0)
}
\arguments{
\item{life_table}{object of class life_table}

\item{x}{x}

\item{t}{t}

\item{m}{m}
}
\description{
probability of death for person at birthday x
}
\examples{
my_table <- data.frame("x" = 0:9,
                       "q_x" = seq(0.05, 0.14, by = 0.01)
                      )

my_table <- life_table(my_table, x = "x", q_x = "q_x")
q_x(my_table, x = 3, t = 5) # probability of x = 3 surviving 5 years
q_x(my_table, x = 2, t = 4, m = 1) # probability of x = 2 surviving 1 year and
# dieing in the following 4 years
}

