% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{qx_data}
\alias{qx_data}
\title{US male and female q_x}
\format{A data frame with 120 observations and 3 variables:
\describe{
\item{x}{exact age}
\item{male_qx}{probability a male dies between birthday x and x + 1}
\item{female_qx}{probability a female dies between birthday x and x + 1}
}}
\usage{
qx_data
}
\description{
Data gathered from the official United States Social Security website 
\url{http://www.ssa.gov/oact/STATS/table4c6.html}
}
\examples{
qx_data
}
\keyword{datasets}

