% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/interest.R
\name{rcir}
\alias{rcir}
\title{Stochastic Cox Ingersol Ross Process}
\usage{
rcir(n, r, b, a, s)
}
\arguments{
\item{n}{number of random observations}

\item{r}{starting short term interest rate}

\item{b}{mean of long term interest rate}

\item{a}{constant that determines rate of reversion of r
to the long term mean b}

\item{s}{volatility of the interest rate process}
}
\description{
Stochastic Cox Ingersol Ross Process
}
\examples{
rcir(n = 10, r = 0.01, b = 0.04, a = 1, s = 0.05)
}

