% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rpv.R
\docType{methods}
\name{rpv,Insuree-method}
\alias{rpv,Insuree-method}
\title{rpv}
\usage{
\S4method{rpv}{Insuree}(object, n, benefit_type = "life")
}
\arguments{
\item{object}{object of class Insuree}

\item{n}{number of observations}

\item{benefit_type}{character string of either "life" or annuity}
}
\description{
rpv
}
\examples{
rdeath(object = Insuree(m_ = 2), n = 5)
rdeath(object = Insuree(x_ = 2, t_ = 3), n = 5)
}

