% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpv.R
\docType{methods}
\name{rpv,Life-method}
\alias{rpv,Life-method}
\title{rpv}
\usage{
\S4method{rpv}{Life}(object, n, interest)
}
\arguments{
\item{object}{object of class Insuree}

\item{n}{number of observations}

\item{interest}{vector of annual interest rates.  Can use the \code{CIR()}
funtion to simulate interest rates in accordance with the Cox Ingersoll Ross
process.}
}
\description{
Simulates the present value of the life insurance benefit for
an object of class \code{Insuree}
}
\examples{
rpv(object = Life(x_ = 2.48, 
                  t_ = 3,
                  m_ = 3,
                  benefit = list(BenefitDeath(t = c(3, 3), value = c(0, 5)))), 
    n = 5,
    interest = Interest(t = 10, rate = 0.04))
    
rpv(object = Life(x_ = 2.48, 
                  t_ = 3,
                  m_ = 0,
                  benefit = list(BenefitAnnuity(t = c(1, 1, 1), value = c(3, 2, 2)))), 
    n = 5,
    interest = Interest(t = 10, rate = 0.04))
    
rpv(object = Life(x_ = 2.48, 
                  t_ = 3,
                  m_ = 0,
                  benefit = list(BenefitAnnuity(t = c(0, 1, 1), value = c(-2, -2, -2)),
                                 BenefitDeath(t = 3, value = 10)
                                 )
                  ), 
    n = 5,
    interest = Interest(t = 10, rate = 0.04))   
}

