% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rpv.R
\docType{methods}
\name{rpv,Pool-method}
\alias{rpv,Pool-method}
\title{rpv}
\usage{
\S4method{rpv}{Pool}(object, n, interest)
}
\arguments{
\item{object}{object of class Insuree}

\item{n}{number of observations}

\item{interest}{vector of annual interest rates.  Can use the \code{CIR()}
funtion to simulate interest rates in accordance with the Cox Ingersoll Ross
process.}
}
\description{
Simulates the present value of the life insurance benefit for
each \code{Insuree} in the a \code{Pool} object.
}
\examples{
rpv(object = Pool(),
    n = 5,
    interest = 0.04)
rpv(object = Pool(),
    n = 5,
    interest = rcir(n = 10, r = 0.01, b = 0.04, a = 1, s = 0.05))
}

