% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rpv.R
\name{rpv}
\alias{rpv}
\title{rpv_life}
\usage{
rpv(object, n, interest)
}
\arguments{
\item{object}{object of class Insuree}

\item{n}{number of observations}

\item{interest}{vector of annual interest rates}
}
\description{
present value of \code{rdeath} simulation
}

