% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rpv.R
\docType{methods}
\name{rpv_annuity,Insuree-method}
\alias{rpv_annuity,Insuree-method}
\title{rpv_annuity}
\usage{
\S4method{rpv_annuity}{Insuree}(object, n)
}
\arguments{
\item{object}{object of class Insuree}

\item{n}{number of observations}
}
\description{
simulate the present value of life contingent
annuity payments for one Insuree
}
\examples{
rpv_annuity(object = Insuree(m_ = .5, benefit = c(1,1, 1, 1)), n = 5)
}

