% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rpv.R
\docType{methods}
\name{rpv_life,Insuree-method}
\alias{rpv_life,Insuree-method}
\title{rpv_life}
\usage{
\S4method{rpv_life}{Insuree}(object, n)
}
\arguments{
\item{object}{object of class Insuree}

\item{n}{number of observations}
}
\description{
Simulates the present value of a single benefit life insurance payment
}
\examples{
rpv_life(object = Insuree(x_ = 2, t_ = 3, benefit_t = c(1, 1, 1), benefit_value = c(3, 2, 1), m_ = 0.3), n = 5)
rpv_life(object = Insuree(x_ = 2.48, t_ = 3.57, benefit_t = c(1, 1, 1, 0.57), benefit_value = c(2, 2, 2, 2), m_ = 3), n = 5)
}

