% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rpv.R
\name{rpv_life}
\alias{rpv_life}
\title{rpv_life}
\usage{
rpv_life(object, n)
}
\arguments{
\item{object}{object of class Insuree}

\item{n}{number of observations}
}
\description{
present value of \code{rdeath} simulation
}

