% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/show.R
\docType{methods}
\name{show,ActuarialTable-method}
\alias{show,ActuarialTable-method}
\title{show ActuarialTable method}
\usage{
\S4method{show}{ActuarialTable}(object)
}
\description{
show ActuarialTable method
}
\examples{
show(ActuarialTable())
}

