% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/show.R
\docType{methods}
\name{show,Insuree-method}
\alias{show,Insuree-method}
\title{show Insuree method}
\usage{
\S4method{show}{Insuree}(object)
}
\description{
show Insuree method
}
\examples{
show(Insuree())
show(Insuree(x_ = 2.5, t_ = 3, benefit = c(1, 1, 4, 1)))
}

