% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sim.R
\docType{methods}
\name{sim,T_x-method}
\alias{sim,T_x-method}
\title{sim}
\usage{
\S4method{sim}{T_x}(object, t_ = NULL)
}
\arguments{
\item{object}{object of class T_x}

\item{t_}{t}
}
\description{
simulated number of complete years for \code{T_x} to survive.  The
simulation is in accordance with the multinomial distribution described
by the life table.
}
\examples{
sim(T_x(), t_ = 3)
}

