% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sim.R
\name{sim}
\alias{sim}
\title{sim}
\usage{
sim(object, t_ = NULL)
}
\description{
simulates \code{T_x} (curtate years until death) in accordance
with the multinomial distribution provided by the life table.
}

