% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\docType{methods}
\name{[,LifeTable,numeric,numeric-method}
\alias{[,LifeTable,numeric,numeric-method}
\title{"["}
\usage{
\S4method{[}{LifeTable,numeric,numeric}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{x}{object of class LifeTable}

\item{i}{x_}

\item{j}{t_}
}
\description{
Helper function for subsetting LifeTable class
}
\details{
The function resturns all `x` and `t` values that are applicable
to the supplied `x_` and `t_` for the individual.  The `q_x` slot is
adjusted for partial years where applicable based on uniform force of
mortality.
}
\examples{
test <- LifeTable()
test[2.5, 0]
test[2.5, 3.5]
test[2, 3]
test[2.4, 3.2]
test[2.4, 3.1]
test[2.2, 0.1]
}

