% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.rpv_Insuree}
\alias{summary.rpv_Insuree}
\title{summary.rpv_Insuree}
\usage{
\method{summary}{rpv_Insuree}(object, ...)
}
\arguments{
\item{object}{\code{rpv_Insuree} object}

\item{...}{other arguments}
}
\description{
summary of rpv() simulation of class \code{Insuree}
}
\examples{
test <- rpv(object = Insuree(),
            n = 500,
            interest = 0.04)
summary(test)
}

