% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.rpv_Life}
\alias{summary.rpv_Life}
\title{summary.rpv_Life}
\usage{
\method{summary}{rpv_Life}(object, ...)
}
\arguments{
\item{object}{\code{rpv_Life} object}

\item{...}{other arguments}
}
\description{
summary of rpv() simulation of class \code{Life}
}
\examples{
test <- rpv(object = Life(benefit = list(BenefitDeath())), 
            n = 500,
            interest = Interest(t = 10, rate = 0.04))
summary(test)
}

