% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LifeTable.R
\docType{class}
\name{T_x-class}
\alias{T_x}
\alias{T_x-class}
\title{T_x}
\description{
An S4 class to represent a life contingent random variable
}
\section{Slots}{

\describe{
\item{\code{x_}}{x value for individual}
}}

