% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{trim_table}
\alias{trim_table}
\title{trim_table}
\usage{
trim_table(table, table_slot, x_ = object[1], t_ = object[length(object)],
  m_ = 0)
}
\arguments{
\item{table}{LifeTable}

\item{table_slot}{slot to be trimmed}

\item{x_}{}

\item{t_}{}

\item{m_}{}
}
\description{
removes all values that are unnecessary for the
specific calculation from the mortaility table column
}
\examples{
trim_table(table = LifeTable(), table_slot = "q_x", x_ = 2, t_ = 3)
}

