% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ActuarialTable.R
\docType{class}
\name{Y_x-class}
\alias{Y_x}
\alias{Y_x-class}
\title{Y_x}
\description{
present value random variable for life contingent
annuity payments
}
\section{Slots}{

\describe{
\item{\code{x_}}{x value for individual}

\item{\code{benefit}}{the annuity benefit}
}}

