% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ActuarialTable.R
\docType{class}
\name{Z_x-class}
\alias{Z_x}
\alias{Z_x-class}
\title{Z_x}
\description{
An S4 class to represent a life contingent random variable
}
\section{Slots}{

\describe{
\item{\code{x_}}{x value for individual}
}}

