% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/interest.R
\name{discount}
\alias{discount}
\title{discount}
\usage{
discount(interest, benefit_time = NA)
}
\arguments{
\item{interest}{vector of annual interest rates}

\item{benefit_time}{the time (from x_) of death}
}
\description{
discount single benefit to present value
}
\examples{
discount(0.04, benefit_time = 1.01)
discount(0.04, benefit_time = 0.8)
}

