% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/discount.R
\name{discount}
\alias{discount}
\title{discount}
\usage{
discount(interest, benefit, ...)
}
\arguments{
\item{interest}{object of class \code{Interest}}

\item{benefit}{object of class \code{Benefit}}

\item{benefit_t}{time benefit is paid}
}
\description{
discounts a life contingent benefit
}

