% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{discount}
\alias{discount}
\title{discount}
\usage{
discount(interest, death_time = NA)
}
\arguments{
\item{interest}{vector of annual interest rates}

\item{death_time}{the time (from x_) of death}
}
\description{
function to discount single benefit
}
\examples{
discount(0.04, death_time = 1.01)
discount(0.04, death_time = 0.8)
}

