% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/expected.R
\docType{methods}
\name{expected,LifeTable-method}
\alias{expected,LifeTable-method}
\title{expected}
\usage{
\S4method{expected}{LifeTable}(object, x_ = object@x[1], t_ = NULL,
  m_ = 0)
}
\arguments{
\item{object}{object of class T_x}

\item{t_}{t}

\item{m_}{m}
}
\description{
expected number of complete years for \code{T_x} to survive
}
\examples{
expected(LifeTable(), x_ = 2, t_ = 3)
}

