% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/expected.R
\docType{methods}
\name{expected,Z_x-method}
\alias{expected,Z_x-method}
\title{expected}
\usage{
\S4method{expected}{Z_x}(object)
}
\arguments{
\item{object}{object of class Z_x}

\item{t_}{t}

\item{m_}{m}
}
\description{
expected present value of an insurance payment \code{Z_x}.  \code{Z_x}
is discounted for interest and mortality.
}
\examples{
# 3 year term insurance with benefit of 4
expected(Z_x(t_ = 3, benefit = 4))
# 3 year term insurance effective in 1 year
expected(Z_x(benefit = 1, x_ = 0, m_ = 1, t_ = 3))
}

