% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/expected.R
\docType{methods}
\name{expected,Z_x-method}
\alias{expected,Z_x-method}
\title{expected}
\usage{
\S4method{expected}{Z_x}(object, t_ = NULL, m_ = 0)
}
\arguments{
\item{object}{object of class Z_x}

\item{t_}{t}

\item{m_}{m}
}
\description{
expected number of complete years for \code{Z_x} to survive discounted for interest
}
\examples{
expected(new("Z_x"), t_ = 3)
}

