% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{hist.rpv_Life}
\alias{hist.rpv_Life}
\title{hist.rpv_Life}
\usage{
\method{hist}{rpv_Life}(object, ...)
}
\arguments{
\item{object}{object of class \code{rpv_Insuree}}

\item{...}{other arguments}
}
\description{
plot a histogram of the present value from a rpv() simulation of the
\code{Life} class.
}
\examples{
test <- rpv(object = Life(benefit = list(BenefitDeath())), 
            n = 500,
            interest = Interest(t = 20, rate = 0.04))
hist(test)
}

