% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{index}
\alias{index}
\title{index}
\usage{
index(object, x_, t_ = 1)
}
\arguments{
\item{object}{LifeTable}

\item{x_}{}

\item{t_}{t_ + m_}
}
\description{
removes all unnecessary values from the LifeTable given the
provided x_, t_, and m_.  For this specific function, t_ + m_
should be provided for the t_ agrument because all information
in the m_ period must be preserved.
}
\details{
The function does not adjust the q_x values for partial years.
It only subsets all values that are relevant to x_, t_, and m_.
See the `[` LifeTable method for partial year q_x values.
}
\examples{
index(LifeTable(), x_ = 2, t_ = 3)
index(LifeTable(), x_ = 2.4, t_ = 3)
index(LifeTable(), x_ = 2.4, t_ = 0)
}

