% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{index}
\alias{index}
\title{index}
\usage{
index(object, x_, t_ = 1)
}
\arguments{
\item{object}{LifeTable}

\item{x_}{}

\item{t_}{t_ + m_}

\item{slot_}{slot to be trimmed}
}
\description{
removes all values that are unnecessary for the
specific calculation from the mortaility table column
}
\examples{
index(LifeTable(), x_ = 2, t_ = 3)
}

