% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Insuree.R
\docType{class}
\name{Insuree-class}
\alias{Insuree}
\alias{Insuree-class}
\title{Insuree}
\description{
An individual with some kind of life contingent insurance
}
\section{Slots}{

\describe{
\item{\code{x_}}{x value for individual}

\item{\code{t_}}{t value for individual}

\item{\code{m_}}{m value for individual}

\item{\code{benefit}}{life contingent benefits.  e.g. For ordinary life insurance
payable at the end of the year of death, the benefit would be a single payment.
For annuity benefits it could be a variable stream of payments.}
}}

