% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LifeTable.R
\docType{class}
\name{LifeTable-class}
\alias{LifeTable}
\alias{LifeTable-class}
\title{LifeTable}
\description{
An S4 class to represent a life table
}
\section{Slots}{

\describe{
\item{\code{x}}{a length n ascending numeric vector}

\item{\code{q_x}}{a length n numeric vector for probability of death at
each element of x}
}}

