% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/p-x.R
\docType{methods}
\name{p_x,LifeTable-method}
\alias{p_x,LifeTable-method}
\title{p_x}
\usage{
\S4method{p_x}{LifeTable}(object, x_ = min(object@x), t_ = 1)
}
\arguments{
\item{object}{object of class LifeTable}

\item{x_}{x_}

\item{t_}{t_}
}
\description{
probability of survival for person at birthday x
}
\examples{
p_x(LifeTable(), x_ = 2.5, t_ = 2)
p_x(LifeTable(), x_ = 2.5, t_ = 2.5)
p_x(LifeTable(), x_ = 2.5, t_ = 3.5)
}

