% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.rpv_life}
\alias{plot.rpv_life}
\title{plot.rpv_life}
\usage{
\method{plot}{rpv_life}(object)
}
\arguments{
\item{object}{rpv_life object}
}
\description{
plot the empirical cumulative distribution of
the present value of death benefits from a rpv_life()
simulation
}
\examples{
test <- rpv_life(object = Insuree(),
                 n = 500,
                 interest = 0.04)
plot(test)
}

