% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.rpv_Pool}
\alias{plot.rpv_Pool}
\title{plot.rpv_Pool}
\usage{
\method{plot}{rpv_Pool}(object, ...)
}
\arguments{
\item{object}{rpv_Pool object}

\item{...}{other arguments}
}
\description{
plot the empirical cumulative distribution of 
the present value of death benefits from a rpv() 
simulation of the \code{Pool} class.
}
\examples{
test <- rpv(object = Pool(), 
            n = 500,
            interest = Interest(t = 5, rate = 0.04))
plot(test)
}

