% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sim.R
\docType{methods}
\name{rdeath,Insuree-method}
\alias{rdeath,Insuree-method}
\title{rdeath}
\usage{
\S4method{rdeath}{Insuree}(object, n)
}
\arguments{
\item{object}{object of class Z_x}

\item{n}{number of observations}
}
\description{
simulated for \code{Z_x}.  The
simulation is in accordance with the multinomial distribution described
by the LifeTable object.
}
\examples{
rdeath(object = Insuree(m_ = 2), n = 5)
rdeath(object = Insuree(x_ = 2.4, t_ = 3, m_ = 0.5, benefit = c(1, 1, 1, 1)), n = 5)
rdeath(object = Insuree(x_ = 3, m_ = 0.2, t_ = 3, benefit = c(1, 1, 1, 1)), n = 5)
}

