% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdeath.R
\docType{methods}
\name{rdeath,Life-method}
\alias{rdeath,Life-method}
\title{rdeath}
\usage{
\S4method{rdeath}{Life}(object, n)
}
\arguments{
\item{object}{object of class Life}

\item{n}{number of observations}
}
\description{
simulated time of death for an object of class \code{Life}.  The
simulation is in accordance with the multinomial distribution described
by the LifeTable object.  Uniform time of death is assumed between x values
indicated by the LifeTable object.
}
\examples{
rdeath(object = Life(m_ = 2), n = 5)
rdeath(object = Life(), n = 1000)
rdeath(object = Life(x_ = 2.4, t_ = 3, m_ = 0.5, benefit = list(BenefitDeath())), n = 5)
max(rdeath(Life(x_ = 2.4, 
                  t_ = 3,
                  m_ = 0), n = 1000)$death_t, na.rm = TRUE)
}

