% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sim.R
\docType{methods}
\name{rdeath,Z_x-method}
\alias{rdeath,Z_x-method}
\title{rdeath}
\usage{
\S4method{rdeath}{Z_x}(object, t_ = (max(object@x) - object@x_), n)
}
\arguments{
\item{object}{object of class Z_x}

\item{t_}{t}

\item{n}{number of observations}
}
\description{
simulated for \code{Z_x}.  The
simulation is in accordance with the multinomial distribution described
by the LifeTable object.
}
\examples{
rdeath(object = Z_x(), t_ = 3, n = 5)
}

