% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rdeath.R
\name{rdeath}
\alias{rdeath}
\title{rdeath}
\usage{
rdeath(object, n)
}
\arguments{
\item{object}{object of class Insuree}

\item{n}{number of observations}
}
\description{
simulates \code{Insuree} time of death in accordance
with the multinomial distribution provided by the life table.
}

