% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sim.R
\name{rdeath}
\alias{rdeath}
\title{rdeath}
\usage{
rdeath(object, t_ = (max(object@x) - object@x_), n)
}
\description{
simulates \code{Insuree} time of death in accordance
with the multinomial distribution provided by the life table.
}

