% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rpv.R
\docType{methods}
\name{rpv,Insuree-method}
\alias{rpv,Insuree-method}
\title{rpv}
\usage{
\S4method{rpv}{Insuree}(object, n, interest)
}
\arguments{
\item{object}{object of class Insuree}

\item{n}{number of observations}

\item{interest}{vector of annual interest rates.  Can use the \code{CIR()}
funtion to simulate interest rates in accordance with the Cox Ingersoll Ross
process.}
}
\description{
Simulates the present value of the life insurance benefit for
an object of class \code{Insuree}
}
\examples{
rpv(object = Insuree(x_ = 2,
                     t_ = 3,
                     benefit_t = c(1, 1, 1),
                     benefit_value = c(3, 2, 1),
                     m_ = 0.3),
    n = 5,
    interest = 0.04)
rpv(object = Insuree(x_ = 2.48,
                     t_ = 3.57,
                     benefit_t = c(1, 1, 1, 0.57),
                     benefit_value = c(2, 2, 4, 2),
                     m_ = 3),
    n = 5,
    interest = rcir(n = 10, r = 0.01, b = 0.04, a = 1, s = 0.05))
}

