% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rpv.R
\docType{methods}
\name{rpv,Insuree-method}
\alias{rpv,Insuree-method}
\title{rpv}
\usage{
\S4method{rpv}{Insuree}(object, n, benefit_type = "life")
}
\arguments{
\item{object}{object of class Insuree}

\item{n}{number of observations}

\item{benefit_type}{character string of either "life" or "annuity"}
}
\description{
rpv
}
\examples{
rpv(object = Insuree(m_ = .5, benefit = c(1,1, 1, 1)), n = 5, benefit_type = "annuity")
rpv(object = Insuree(x_ = 2, t_ = 3, benefit = c(1, 1, 1, 1), m_ = 0.3), n = 5)
}

