% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rpv.R
\docType{methods}
\name{rpv_annuity,Insuree-method}
\alias{rpv_annuity,Insuree-method}
\title{rpv_annuity}
\usage{
\S4method{rpv_annuity}{Insuree}(object, n)
}
\arguments{
\item{object}{object of class Insuree}

\item{n}{number of observations}
}
\description{
simulates the present value of life contingent
annuity payments for one \code{Insuree}
}
\examples{
rpv_annuity(object = Insuree(x_ = 2.48,
                             t_ = 3.57,
                             benefit_t = c(1, 1, 1, 0.57),
                             benefit_value = c(2, 2, 4, 2),
                             m_ = 3),
            n = 5)
}

