% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\docType{methods}
\name{[,LifeTable,numeric,numeric-method}
\alias{[,LifeTable,numeric,numeric-method}
\title{"["}
\usage{
\S4method{[}{LifeTable,numeric,numeric}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{x}{object of class ActuarialTable}

\item{i}{x_}

\item{j}{t_}
}
\description{
Helper function for subsetting ActuarialTables
I want to use this to replace the trim_table() function
}
\details{
Need to expand so that it works for LifeTable before removing trim_table
Also need to add
}
\examples{
test <- LifeTable()
test[2.5, 0]
test[2.5, 3.5]
test[2, 3]
}

