% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.rpv_life}
\alias{summary.rpv_life}
\title{summary.rpv_life}
\usage{
\method{summary}{rpv_life}(object, ...)
}
\arguments{
\item{object}{rpv_life object}

\item{...}{other arguments}
}
\description{
summary of rpv_life() simulation
}
\examples{
test <- rpv_life(object = Insuree(),
                 n = 500,
                 interest = 0.04)
summary(test)
}

