% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{tp_x8q_x}
\alias{tp_x8q_x}
\title{tp_x8q_x}
\usage{
tp_x8q_x(object)
}
\arguments{
\item{object}{Insuree object}
}
\description{
returns the probability of death in each x for a percon age x_
}
\examples{
tp_x8q_x(Insuree(x_ = 2, t_ = 3))
tp_x8q_x(Insuree(x_ = 2.4, t_ = 3))
}

