% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{tp_x8q_x}
\alias{tp_x8q_x}
\title{tp_x8q_x}
\usage{
tp_x8q_x(object, x_, t_ = NULL, m_ = 0)
}
\arguments{
\item{object}{LifeTable object}

\item{x_}{}

\item{t_}{}

\item{m_}{}
}
\description{
returns the probability of death in each x for a percon age x_.  This function
converts the `q_x` slot from the probability of death given the individual
is age x to the probability of death given the individual is age x_.
}
\examples{
tp_x8q_x(LifeTable(), x_ = 2, t_ = 3, m_ = 0)
tp_x8q_x(LifeTable(), x_ = 2.4, t_ = 3, m_ = 0.5)
tp_x8q_x(LifeTable(), x_ = 3, m_ = 0.2, t_ = 3)
}

