% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{trim_table}
\alias{trim_table}
\title{trim_table}
\usage{
trim_table(object, x_ = object@x[1], t_ = NULL, m_ = 0)
}
\arguments{
\item{object}{LifeTable}

\item{x_}{x_}

\item{t_}{t_}

\item{m_}{m_}
}
\description{
Helper function for creating new LifeTables given the
x_, t_, and m_ values.
}
\examples{
trim_table(LifeTable(), x_ = 2, t_ = 3, m_ = 1)
}

