% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ActuarialTable.R
\docType{class}
\name{Z_x-class}
\alias{Z_x}
\alias{Z_x-class}
\title{Z_x}
\description{
The present value random variable for a life contingent
insurance payment payable at the end of the year of death
}
\section{Slots}{

\describe{
\item{\code{x_}}{x value for individual}

\item{\code{benefit}}{the amount payable at the end of the year of death}
}}

